// client/other/role-list.ts
export enum RoleList {
    ALL = "ALL",
    ADMIN = "ADMIN",
    ALL_INFO_USERS = "ALL_INFO_USERS",
    PUBLICATIONS = "PUBLICATIONS",
    VERIFIED = "VERIFIED",
    UNIVERSITY = "UNIVERSITY",
    USER = "USER",
    USER_CONTROL = "USER_CONTROL",
    USER_ROLE = "USER_ROLE",
    USER_POSITION = "USER_POSITION",
    USER_COMMENT = "USER_COMMENT"
}